v2.5: Stappler fixes/refactoring
Common:
- NetworkHandle, Bitmap refactored to use StringView instead of String

Layout:
- MovingAverage: signed-unsigned warning fixed
- Canvas: fixed batching ending bug

Stappler:
- Network, Asset, TextureCache and Thread refactored to use StringView
- draw::PathNode fixed missed second buffer/offscreen texture

Serenity:
- storage: File and Image fields now can be updated with integers (oids)
- storage::File::createFile - data overload with memory data

+++ layout/types/SLMovingAverage.h
@@ -31,7 +31,7 @@ template <size_t Count>
 class MovingAverage {
 public:
     void dropValues() {
-        for (int i = 0; i < Count; i++) {
+        for (size_t i = 0; i < Count; i++) {
             _values[i] = 0;
         }
     }
